// packages/features/worker-runtime/src/integrations/framework/types.ts
export type IntegrationCommandExecute = (input: {
  arguments: Record<string, unknown>;
  context: IntegrationExecutionContext;
}) => Promise<unknown>;
// ...
export type IntegrationExecutionContext = {
  auth: ConnectedOauthAccessRecord | null;
  tenantIntegrationId: string | null;
};